Refactor components, add tests and docs
